requirements.txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.42.0
langchain_huggingface
faiss-cpu